// src/app/models/usersignin.ts
export class UserSignIn {
    username?: string;
    password?: string;
    nomeCompleto?: string;
    matricula?: string;
    cpf?: string;
    rg?: string;
    orgaoEmissor?: string;
    tamanhoCamisa?: string;
    email?: string;
    telefone?: string;
    primeiraGrad?: boolean;
    dataEntrada?: string;//Tipo data
}